Change UIView to UIControl
